+++ tests_my/pages/dashboard/login_page.py
@@ -13,6 +13,7 @@ class LoginPage():
         self._cbb_repository = ComboBox("id=repository")
 
     def fill_info_login(self, username, password, repository):
+        self._cbb_repository.wait_for_visible()
         if repository is None:
             self._cbb_repository.select_by_visible_text(constants.REPOSITORY_DEFAULT)
         else:
@@ -20,6 +21,10 @@ class LoginPage():
         self._txt_username.send_keys(username)
         self._txt_password.send_keys(password)
 
-    def login(self, username, password, repository=None):
+    def login(self, username=constants.USER_NAME, password=constants.PASSWORD, repository=None):
         self.fill_info_login(username, password, repository)
+        self.submit_login()
+
+    def submit_login(self):
+        self._btn_login.wait_for_visible()
         self._btn_login.click()
